+++ tests/installation_check.py
@@ -26,7 +26,7 @@ try:
 
     # Basic package usage
     fl_obj = FluorObj(typical_size=2.0); fl_obj.get_shape(accelerated=True)
-    scene = UscopeScene(width=14, height=12); obj_pl = scene.set_random_places([fl_obj])
+    scene = UscopeScene(width=14, height=12); obj_pl = scene.set_random_places(tuple([fl_obj]))
     scene.put_objects_on(obj_pl)
 
 except (ImportError, ModuleNotFoundError):